feat: add Azure Functions integration and deployment workflows
Introduce Azure Functions support with HTTP trigger and deployment setup.
