+++ src/functions/http-trigger.ts
@@ -1,0 +1,10 @@
+import { app } from "@azure/functions";
+import { azureHonoHandler } from "@marplex/hono-azurefunc-adapter";
+import { app as honoApp } from "../worker";
+
+app.http("http-trigger", {
+  methods: ["GET", "POST"],
+  authLevel: "anonymous",
+  route: "{*proxy}",
+  handler: azureHonoHandler(honoApp.fetch),
+});
